VaTreeView story rework
